pyproject.toml
[project]
name = "browser-agent-nvidia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "click>=8.3.1",
    "einops>=0.8.2",
    "numpy>=2.2.6",
    "open-clip-torch>=3.2.0",
    "pillow>=12.1.0",
    "playwright>=1.58.0",
    "pyyaml>=6.0.3",
    "timm>=1.0.24",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=4.46.0,<5.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
